Update aboutme.md
Removed: Who am I?:
- A keen Liverpool FC fan, hailing from the North West of England
- An adopted San Francisco 49ers fan, inspired by the 2013 team of Smith and Kaepernick, following the highs and lows since.

This website is here to showcase what I do outside of work. I'm passionate about data, and specifically how it can be applied in sports, particularly football (soccer) and American Football. At the time of writing, advanced analytics in football is a burgeoning area which is beginning to catch up to the levels of usage of analytics in American sports. I want to demonstrate my understanding of analytics on American Football data, and demonstrate how similar advances can be made in football.

Ultimately I want this blog to show:
- My capability in sports analytics, delivering some interesting insights where possible
- My passion for data viz, with some great examples of how effective visualisation can be the slam dunk you need to demonstrating an idea
- My journey of self-education - I'm very interested in sports analytics, but I recognise that my skills are very much in their formative years!